Fix line highlighting in ibuffer menu.
Turning on hl-line-mode in an ibuffer buffer, when also using
ibuffer-auto-mode does not work in that the initial line highlighting
will not appear (it will the very first time ibuffer is activated, but
not on the following activations).

This patch fixes this by adding an advice to the ibuffer redisplay
engine.