git: Move attributes file
My .gitattributes file in the home directory was not being used.  I
don't know if this used to be the right place, but no longer is, or if I
was jut always wrong...

The new location appears to be correct.  For safety, I have, for now,
added a symlink from the old location to the new location, in case older
versions of git did use the location in the home directory.