emacs: New way to store notmuch-saved-searches
This commit adds a new functions that extends the notmuch-saved-searches
variable by loading additional searches from a new file in my home
directory.

This change allows me to store private searches into a different file
that I don't commit to git.  This file will include searches that
contain things like customer names, that are confidential.

The additional searches are loaded when the notmuch package is loaded,
and appended to the notmuch-saved-searches list.