bash: Start using module system for managing tools
The module system allows tools to be inserted and removed from the
environment and can set different environment variables as needed.

With this commit I now build and install tools into the directory
~/.software/packages with a name and version prefix, so, for example,

  ~/.software/packages/git/2.2.2/

would be where I installed version 2.2.2 of git.

Then I create a modulefile for this package like this:

  ~/.software/modules/git/2.2.2

now I can do:

  $ module load git/2.2.2

and I have all of the environment settings I need.

For simplicity, all of the tool modulefiles just load a single common
module file that does all of the actual work.

As I install tools into the packages directory I might loose track and
forget to create the required modulefiles, so I have a script
module-sync, that will find and create module files as needed.