Better support for symlinks already existing
The perl '-e' check returns true or false based on the target of a
symlink, so, a broken link will report as not existing, which was
leading the script to try and recreate the symlink.

Changed so that we first check the '-l' (a symlink), then check for
'-e' (anything else existing), before finally, trying to create a new
symlink.