git: Replace 'brr' command with 'bgrep'
The new 'git bgrep' command is a more generic version of the 'git brr'
command, this commit makes 'brr' an alias for 'bgrep'.

The new bgrep command takes this form:

   git bgrep [<remote>] <pattern>
   git bgrep -a <pattern>

If <remote> is given then the searching is performed in the named
remote, if '-a' is given then all remotes are searched.

The searching is for branches matching <pattern>.  Any branches that
match <pattern> are listed.

The patterns have an implicit '*' added to the end, but not to the
start, so 'gdb-' will find branches that match 'gdb-*' (using glob
style matchinng).

It is possible to supply a pattern like '*gdb*' to match any branch
name that contains the text 'gdb'.