git: Rewrite bgrep command in perl, add brr command
This is a complete rewrite of the bgrep command in perl, with the goal
of making this more maintainable.  The biggest difference is that the
branch patterns are now perl regexps rather than shell globs.

One area where this rewrite works that the old shell version didn't is
for finding branches that are nested within a namespace, such as:

  remote-name/namespace/branch-name

which previously would not have been matched.

The 'brr' command is just an alias that I've had around for a while on
this matching that lists all of the branches (by default local
branches), and the branches will be listed by last commit date (like
bgrep).