emacs/org: Only initialise org-agenda-files if not already set
When trying to view personalised agenda views that pull their contents
fromm different org files I ran into the problem that when the
visualisation buffer was put into org-mode, the org-agenda-files list
would be overritten back to the tasks.org list, and the agenda
contents would be lost.

To prevent this, only reset org-agenda-files on entering org-mode if
this is not already set.