bash: Update for the upto_core.pl helper script
This script is used by the bash 'upto' command.  This commit just
rewrites it a bit, adds proper command line argument processing, and
adds support for listing possible completions.  Ideally this will (in
future) be combined with a bash completer so that the 'upto' command
will give command line completion.  Maybe.... one day....

The bashrc file is also updated to use the new flags when running the
upto_core.pl script.