i3/toggl: Improve performance of toggle project switching script
Improve the performance of the script for switching toggl project.
Cache the list of projects as this changes very infrequently, this
allows the selection dialog to open much quicker.

Each time this is run, in the background, the project list file is
updated.

This means that if the project list _is_ changed, then invoke this
script, the new project will be missing.  Then invoke the script
again, and the new project will now be in place.