toggl: Allow user to quit out of project selector
Passing `-only-match` to rofi means that the user can neither enter a
custom response, nor can they quit with <escape>, they are forced to
make a selection.

This is not what I want, instead use `-no-custom`, this prevents
custom replies, but does allow the user to abort with <escape>, in
which case the empty string is returned.