modules: Add support for per-package .modulerc files
With this commit there is now support for two per-package modulerc
files, in these locations:

   <prefix>/packages/<package-name>/<version-number>/.modulerc
   <prefix>/packages/<package-name>/.modulerc

The files are checked for in this order, and only the first file found
is sourced.

The comments in module-common describe more about what these files can
be used for and how, but basically it allows for per-package tweaking
of the environment setup.