desktop: smart screensaver disable script
A script that can be used by xfce desktop environment to keep the
screensaver from activating.  The benefit of this script is that if I
manually activate the screensaver then the script does not "poke" the
screensaver like the default does.

The default script always pokes the screensaver every few seconds, and
if the screensaver is already active then this causes the password
prompt to appear, and prevents the screensaver ever putting the
display to sleep.

This script avoids this problem.